Move values to their own module
