+++ tokio-trace/src/lib.rs
@@ -113,7 +113,7 @@
 
 extern crate futures;
 
-use std::{any::Any, fmt, slice};
+use std::{fmt, slice};
 
 #[doc(hidden)]
 #[macro_export]
@@ -292,45 +292,14 @@ pub enum Level {
 mod dispatcher;
 pub mod span;
 pub mod subscriber;
+pub mod value;
 
 pub use self::{
     dispatcher::Dispatch,
     span::{Data as SpanData, Id as SpanId, Span},
     subscriber::Subscriber,
+    value::*,
 };
-
-// XXX: im using fmt::Debug for prototyping purposes, it should probably leave.
-pub trait Value: fmt::Debug + Send + Sync { }
-
-pub trait OwnedValue: Value + Any {
-}
-
-impl<T> Value for T
-where
-    T: fmt::Debug + Send + Sync,
-{
-
-}
-
-pub trait Duplicate: Value {
-    // like `Clone`, but "different"
-    fn duplicate(&self) -> Box<dyn OwnedValue>;
-}
-
-impl<T> OwnedValue for T
-where
-    T: Any + Value,
-{ }
-
-impl<T> Duplicate for T
-where
-    T: Value + ToOwned,
-    <T as ToOwned>::Owned: OwnedValue,
-{
-    fn duplicate(&self) -> Box<dyn OwnedValue> {
-        Box::new(self.to_owned())
-    }
-}
 
 /// **Note**: `Event` must be generic over two lifetimes, that of `Event` itself
 /// (the `'event` lifetime) *and* the lifetime of the event's metadata (the
